One Front Door/SourceCode/DPSG.Web.Merchandiser/src/WebApp/src/app/dispatch/route-item/route-item.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { RouteItemComponent } from './route-item.component';

describe('Component: RouteItem', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [RouteItemComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([RouteItemComponent],
      (component: RouteItemComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(RouteItemComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(RouteItemComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-route-item></app-route-item>
  `,
  directives: [RouteItemComponent]
})
class RouteItemComponentTestController {
}

